requirements.txt
# 动火离人安全监测系统 - 依赖包

# 核心框架
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# 计算机视觉
opencv-python>=4.9.0
numpy>=1.24.0

# YOLO推理
ultralytics>=8.1.0

# 配置管理
pyyaml>=6.0

# 语音播报
pyttsx3>=2.90

# WebSocket
websockets>=12.0

# 开发工具
python-multipart>=0.0.6
